# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A comprehensive project management backend API for DevHive",
        "title": "DevHive Backend API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "DevHive Team",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/mail/send": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Sends an email to the specified recipient using the mail service",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mail"
                ],
                "summary": "Send email",
                "parameters": [
                    {
                        "description": "Email request containing To, Subject, and Body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.EmailRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Email sent successfully",
                        "schema": {
                            "$ref": "#/definitions/models.EmailResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request - invalid email request",
                        "schema": {
                            "$ref": "#/definitions/models.EmailError"
                        }
                    },
                    "500": {
                        "description": "Internal server error - email sending failed",
                        "schema": {
                            "$ref": "#/definitions/models.EmailError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.EmailError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {
                    "type": "string"
                },
                "error": {
                    "type": "string"
                }
            }
        },
        "models.EmailRequest": {
            "type": "object",
            "required": [
                "body",
                "subject",
                "to"
            ],
            "properties": {
                "body": {
                    "type": "string",
                    "example": "This is the email content"
                },
                "subject": {
                    "type": "string",
                    "example": "Important Notification"
                },
                "to": {
                    "type": "string",
                    "example": "recipient@example.com"
                }
            }
        },
        "models.EmailResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Email sent successfully!"
                },
                "sentAt": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}